Add config types
