--- src/config/configError.ts
@@ -1,6 +1,0 @@
-export class ConfigError extends Error {
-  constructor(message: string) {
-    super(message);
-    this.name = "ConfigError";
-  }
-}
